refactor: ♻️ refactor redux async
